LPD-45147 Forward compatibility, avoid directly implements HttpServletRequest
