+++ modules/apps/portal-security/portal-security-service-access-policy-service/src/main/java/com/liferay/portal/security/service/access/policy/internal/SAPAccessControlPolicy.java
@@ -293,7 +293,7 @@ public class SAPAccessControlPolicy extends BaseAccessControlPolicy {
 			if (GetterUtil.getBoolean(
 					httpServletRequest.getAttribute(
 						"com.liferay.portal.vulcan.internal.template.servlet." +
-							"RESTClientHttpRequest"))) {
+							"RESTClientHttpRequestDelegate"))) {
 
 				systemServiceAccessPolicyNames.add(
 					sapConfiguration.
