+++ modules/apps/portal-vulcan/portal-vulcan-impl/src/main/java/com/liferay/portal/vulcan/internal/template/RESTClientTemplateContextContributor.java
@@ -17,7 +17,8 @@ import com.liferay.portal.kernel.servlet.ServletContextPool;
 import com.liferay.portal.kernel.template.TemplateContextContributor;
 import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.Portal;
-import com.liferay.portal.vulcan.internal.template.servlet.RESTClientHttpRequest;
+import com.liferay.portal.kernel.util.ProxyUtil;
+import com.liferay.portal.vulcan.internal.template.servlet.RESTClientHttpRequestDelegate;
 import com.liferay.portal.vulcan.internal.template.servlet.RESTClientHttpResponse;
 
 import java.util.Map;
@@ -89,8 +90,12 @@ public class RESTClientTemplateContextContributor
 				AccessControlUtil.setAccessControlContext(null);
 
 				requestDispatcher.forward(
-					new RESTClientHttpRequest(
-						_contextObjects, _httpServletRequest, path),
+					ProxyUtil.newDelegateProxyInstance(
+						HttpServletRequest.class.getClassLoader(),
+						HttpServletRequest.class,
+						new RESTClientHttpRequestDelegate(
+							_contextObjects, _httpServletRequest, path),
+						_httpServletRequest),
 					httpServletResponse);
 			}
 			finally {
